Refactored tests to work with the new tensor class.
