+++ src/base/ITensor.ts
@@ -11,7 +11,7 @@ export default interface ITensor<TensorType> {
     print_info: (title: string) => void;
     
     // get_axis_iterable: (n: number) => Generator<TensorType>;
-    //
+    
     // // init operations
     zeros: () => TensorType;
     ones: () => TensorType;
@@ -19,13 +19,13 @@ export default interface ITensor<TensorType> {
     rand: (min: number, max: number) => TensorType;
     rand_int: (min: number, max: number) => TensorType;
     fill: (value: number) => TensorType;
-    //
+    
     // // memory dealloc operation
     // free: () => void;
-    //
+    
     // // view operations
     // create_view: (axis: number, offset: number) => TensorType;
-    // transpose: (...permutation: number[]) => TensorType;
+    transpose: (...permutation: number[]) => TensorType;
 
     // unary operations
     relu:       () => TensorType;
@@ -68,7 +68,7 @@ export default interface ITensor<TensorType> {
     mean: () => TensorType;
 
     // operation shorthands
-    // readonly T: TensorType;
+    readonly T: TensorType;
 
     // [Symbol.iterator](): Iterator<TensorType>;
 }
